Moved GithubUser onto entity dir
