+++ src/entity/GithubUser.js
